api/requirements.txt
azure-functions
requests
python-dotenv
azure-data-tables
